fix(auth): relocate project folders for better structure
