+++ src/main/java/com/onspring/onspring_customer/domain/auth/dto/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.onspring.onspring_customer.global.auth.dto;
+package com.onspring.onspring_customer.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,10 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class TokenResponseDto {
+public class LoginResponseDto {
     private String accessToken;
+    private String refreshToken;
+    private Long id;
+
+
 }
